move generic logic to hook
