+++ src/main/java/at/hannibal2/skyhanni/features/garden/GardenWarpCommands.kt
@@ -4,7 +4,7 @@ import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.data.IslandType
 import at.hannibal2.skyhanni.events.MessageSendToServerEvent
 import at.hannibal2.skyhanni.events.minecraft.KeyDownEvent
-import at.hannibal2.skyhanni.features.misc.MouseSensitivityManager
+import at.hannibal2.skyhanni.features.misc.LockMouseLook
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.ChatUtils
 import at.hannibal2.skyhanni.utils.HypixelCommands
@@ -46,14 +46,14 @@ object GardenWarpCommands {
         if (message == "/barn") {
             event.cancel()
             HypixelCommands.teleportToPlot("barn")
-            MouseSensitivityManager.unlockMouse()
+            LockMouseLook.unlockMouse()
         }
 
         tpPlotPattern.matchMatcher(event.message) {
             event.cancel()
             val plotName = group("plot")
             HypixelCommands.teleportToPlot(plotName)
-            MouseSensitivityManager.unlockMouse()
+            LockMouseLook.unlockMouse()
         }
     }
 
@@ -78,7 +78,7 @@ object GardenWarpCommands {
                 if (lastWarpTime.passedSince() < 2.seconds) return
                 lastWarpTime = SimpleTimeMark.now()
 
-                MouseSensitivityManager.unlockMouse()
+                LockMouseLook.unlockMouse()
                 HypixelCommands.teleportToPlot("barn")
             }
